Refactor folder division and added final step: load in differents formats
